+++ src/renderer/i18n/types.ts
@@ -117,16 +117,11 @@ export type CommonMessageKey =
   | 'common.swap'
   | 'common.chainDisabled'
   | 'common.savers'
-  | 'common.lending'
-  | 'common.borrow'
-  | 'common.repay'
   | 'common.earn'
   | 'common.earnings'
   | 'common.allocationByType'
   | 'common.allocationByChain'
   | 'common.liquidity'
-  | 'common.collateral'
-  | 'common.debt'
   | 'common.withdraw'
   | 'common.approve'
   | 'common.accept'
@@ -630,36 +625,6 @@ type RunePoolMessageKey =
 
 export type RunePoolMessages = { [key in RunePoolMessageKey]: string }
 
-type LoanMessageKey =
-  | 'loan.noLoans'
-  | 'loan.openLoan'
-  | 'loan.closeLoan'
-  | 'loan.detail.title'
-  | 'loan.detail.debt.current'
-  | 'loan.detail.debt.issued'
-  | 'loan.detail.collateral.deposited'
-  | 'loan.detail.collateral.current'
-  | 'loan.detail.collateral.withdrawn'
-  | 'loan.detail.age'
-  | 'loan.detail.lastRepay'
-  | 'loan.detail.repayed'
-  | 'loan.detail.assetAmount'
-  | 'loan.detail.collaterizationRatio'
-  | 'loan.info.max.loan.value'
-  | 'loan.info.max.balance'
-  | 'loan.borrow.state.sending'
-  | 'loan.borrow.state.checkResults'
-  | 'loan.borrow.state.pending'
-  | 'loan.borrow.state.success'
-  | 'loan.borrow.state.error'
-  | 'loan.repay.state.sending'
-  | 'loan.repay.state.checkResults'
-  | 'loan.repay.state.pending'
-  | 'loan.repay.state.success'
-  | 'loan.repay.state.error'
-
-export type LoanMessages = { [key in LoanMessageKey]: string }
-
 export type HaltMessageKey =
   | 'halt.thorchain'
   | 'halt.trading'
@@ -681,7 +646,6 @@ export type Messages = CommonMessages &
   DepositMessages &
   SaversMessages &
   RunePoolMessages &
-  LoanMessages &
   LedgerMessages &
   BondsMessages &
   PoolSharesMessage &