+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Add 
 - min threshold dust amount check [#590](https://github.com/asgardex/asgardex-desktop/pull/591)
 
+## Updates/Fixes
+- removed all loan/lending logic [#603](https://github.com/asgardex/asgardex-desktop/pull/603)
+
 # 1.30.1 (2025-1-17)
 
 ## add